api-service/requirements.txt
flask==2.0.1
werkzeug==2.0.3
kafka-python==2.0.2
requests==2.28.1
pytest==7.4.0